converted to bulma css framework, removed bootstrap, removed jquery

+++ resources/views/admin/article/edit.blade.php
@@ -16,33 +16,31 @@
 
 
 @section('heading')
-    <span class="fa fa-pencil-square-o"></span> Edit Article
+    <span class="icon is-medium"><i class="fa fa-pencil-square-o"></i></span> Edit Article
+    @can('update-article')
+        <span class="pull-right">
+            {!! Form::model($article, ['id'=>'delete-form','method' => 'DELETE', 'files' => true, 'action' => ['Admin\ArticleController@destroy',$article]]) !!}
+            {!! Form::submit('Delete',['class'=>'button is-danger']) !!}
+            {!! Form::close() !!}
+        </span>
+    @endcan
 @stop
 
 
 @section('content')
 
-    <div class="row">
-        <div class="col-sm-12">
-            {!! Form::model($article, ['id'=>'edit-form','method' => 'PATCH', 'files' => true, 'action' => ['Admin\ArticleController@update',$article]]) !!}
+    <div class="container">
+        {!! Form::model($article, ['id'=>'edit-form','method' => 'PATCH', 'files' => true, 'action' => ['Admin\ArticleController@update',$article]]) !!}
 
-            @include('admin.article.forms.info')
+        @include('admin.article.forms.info')
 
-            @include('admin.article.forms.body')
+        @include('admin.article.forms.body')
 
-            @include('admin.article.forms.image')
+        @include('admin.article.forms.image')
 
-            {!! Form::submit('Save',['class'=>'btn btn-primary btn-block btn-lg']) !!}
-            {!! Form::close() !!}
-        </div>
-    </div>
-
-    <div class="row" style="margin-top:20px;">
-        <div class="col-sm-12">
-            {!! Form::model($article, ['id'=>'delete-form','method' => 'DELETE', 'files' => true, 'action' => ['Admin\ArticleController@destroy',$article]]) !!}
-            {!! Form::submit('Delete',['class'=>'btn btn-danger btn-block btn-lg']) !!}
-            {!! Form::close() !!}
-        </div>
+        {!! Form::submit('Save',['class'=>'button is-primary is-large']) !!}
+        <a class="button is-large" href="{{ URL::previous() }}"> Back </a>
+        {!! Form::close() !!}
     </div>
 
 @endsection
