+++ resources/views/admin/category/edit.blade.php
@@ -16,29 +16,27 @@
 
 
 @section('heading')
-    <span class="fa fa-pencil-square-o"></span> Edit Category
+    <span class="icon is-medium"><i class="fa fa-pencil-square-o"></i></span> Edit Category
+    @can('update-category')
+        <span class="pull-right">
+            {!! Form::model($category, ['id'=>'delete-form','method' => 'DELETE', 'files' => true, 'action' => ['Admin\CategoryController@destroy',$category]]) !!}
+            {!! Form::submit('Delete',['class'=>'button is-danger']) !!}
+            {!! Form::close() !!}
+        </span>
+    @endcan
 @stop
 
 
 @section('content')
 
-    <div class="row">
-        <div class="col-sm-12">
-            {!! Form::model($category, ['id'=>'edit-form','method' => 'PATCH', 'files' => true, 'action' => ['Admin\CategoryController@update',$category]]) !!}
+    <div class="container">
+        {!! Form::model($category, ['id'=>'edit-form','method' => 'PATCH', 'files' => true, 'action' => ['Admin\CategoryController@update',$category]]) !!}
 
-            @include('admin.category.form')
+        @include('admin.category.form')
 
-            {!! Form::submit('Save',['class'=>'btn btn-primary btn-block btn-lg']) !!}
-            {!! Form::close() !!}
-        </div>
-    </div>
-
-    <div class="row" style="margin-top:20px;">
-        <div class="col-sm-12">
-            {!! Form::model($category, ['id'=>'delete-form','method' => 'DELETE', 'files' => true, 'action' => ['Admin\CategoryController@destroy',$category]]) !!}
-            {!! Form::submit('Delete',['class'=>'btn btn-danger btn-block btn-lg']) !!}
-            {!! Form::close() !!}
-        </div>
+        {!! Form::submit('Save',['class'=>'button is-primary is-large']) !!}
+        <a class="button is-large" href="{{ URL::previous() }}"> Back </a>
+        {!! Form::close() !!}
     </div>
 
 @endsection
